% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples2.R
\name{ARIMASearch.BIC}
\alias{ARIMASearch.BIC}
\title{Example2}
\usage{
ARIMASearch.BIC(chromosome, plen = 2, XMat, Xt)
}
\arguments{
\item{chromosome}{The chromosome consists of the number of changepoints, the
order of AR part, the order of MA part, the changepoint locations, and a
value of time series length plus 1 indicating the end of the chromosome.}

\item{plen}{Since we want to perform the model order selection and the changepoint
detection task, \code{plen} is nonzero and is set as the number of model
order parameter, which is 2.}

\item{XMat}{A matrix contains the covariates (not includes changepoint effects)
for time series regression.}

\item{Xt}{The simulated ARMA(2,1) time series from \code{ts.sim} function.}
}
\value{
Returned the value of the obejctive function (i.e. BIC).
}
\description{
The example objective function for changepoint search in Autoregressive
moving average (ARMA(2,1)) via Bayesian Information Criterion (BIC).
}
\examples{
Ts = 1000
betaT = c(0.5) # intercept
XMatT = matrix(1, nrow=Ts, ncol=1)
colnames(XMatT) = "intercept"
sigmaT = 1
phiT = c(0.5)
thetaT = NULL
DeltaT = c(2, -2)
Cp.prop = c(1/4, 3/4)
CpLocT = floor(Ts*Cp.prop)

myts = ts.sim(beta=betaT, XMat=XMatT, sigma=sigmaT, phi=phiT, theta=thetaT,
              Delta=DeltaT, CpLoc=CpLocT, seed=1234)

# candidate changepoint configuration
chromosome = c(2, 2, 1, 250, 750, 1001)
ARIMASearch.BIC(chromosome, plen=2, XMat=XMatT, Xt=myts)
}
