% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayaCGH.R
\docType{data}
\name{arrayaCGH}
\alias{arrayaCGH}
\title{Glioblastoma array aCGH data}
\format{
An object of class \code{data.frame} with 193 rows and 5 columns.
}
\usage{
data(arrayaCGH)
}
\description{
Glioblastoma multiforme (GBM) can cause chromosomal aberrations during DNA
replication. Array comparative genomic hybridization (aCGH) experiments analyze
these aberrations using fluorescence intensity ratios, which indicate DNA copy
number variations.
}
\examples{
data(arrayaCGH)
}
\references{
Lai, W. R., Johnson, M. D., Kucherlapati, R., & Park, P. J. (2005).
Comparative analysis of algorithms for identifying amplifications and deletions
in array CGH data. \emph{Bioinformatics}, 21(19), 3763-3770.

Killick, R., & Eckley, I. A. (2014). changepoint: An R package for
changepoint analysis. \emph{Journal of statistical software}, 58, 1-19.
}
\keyword{datasets}
