% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples1.R
\name{BinSearch.BIC}
\alias{BinSearch.BIC}
\title{Example1}
\usage{
BinSearch.BIC(chromosome, plen = 0, Xt)
}
\arguments{
\item{chromosome}{The chromosome consists of the number of changepoints, their
locations, and a value of time series length plus 1 indicating the end of the
chromosome.}

\item{plen}{Since we only want to perform the changepoint detection task,
\code{plen} has to be set as 0.}

\item{Xt}{The simulated AR(1) time series from \code{ts.sim} function.}
}
\value{
Returned the value of the obejctive function (i.e. BIC).
}
\description{
The example objective function for changepoint search in First order Gaussian
autoregressions (AR(1)) via Bayesian Information Criterion (BIC).
}
\details{
The AR order here is fixed to be 1 and we only want to conduct changepoint
detection in this example.
}
\examples{
Ts = 1000
betaT = c(0.5) # intercept
XMatT = matrix(1, nrow=Ts, ncol=1)
colnames(XMatT) = "intercept"
sigmaT = 1
phiT = c(0.5)
thetaT = NULL
DeltaT = c(2, -2)
Cp.prop = c(1/4, 3/4)
CpLocT = floor(Ts*Cp.prop)

myts = ts.sim(beta=betaT, XMat=XMatT, sigma=sigmaT, phi=phiT, theta=thetaT,
              Delta=DeltaT, CpLoc=CpLocT, seed=1234)

# candidate changepoint configuration
chromosome = c(2, 250, 750, 1001)
BinSearch.BIC(chromosome, Xt=myts)
}
