% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cptDist.R
\name{cptDist}
\alias{cptDist}
\title{Comparing multiple changepoint configurations by pairwise distance}
\usage{
cptDist(tau1, tau2, n)
}
\arguments{
\item{tau1}{A vector contains the changepoint locations for \eqn{C_{1}}
for comparison. A value \code{NULL} is required if there is no changepoint detected.}

\item{tau2}{A vector contains the changepoint locations for  \eqn{C_{2}}
for comparison. A value \code{NULL} is required if there is no changepoint detected.}

\item{n}{The simulated time series sample size. Two changepoint configurations
should have the same \eqn{n} values.}
}
\value{
\item{dist}{The calculated distance.}
}
\description{
Given a true changepoint configuration, \eqn{C_{1}=\{\tau_{1}, \ldots, \tau_{m}\}},
for time series simulation, we want to compare an estimated multiple
changepoint configuration, \eqn{C_{2}=\{\eta_{1}, \ldots, \eta_{k}\}}, to \eqn{C_{1}} to
examine the detection algorithm performance.
}
\details{
The pairwise distance was proposed by Shi et al. (2022),
\deqn{d(C_{1}, C_{2})=|m-k|+ min(A(C_{1}, C_{2})),}
where \eqn{m} is the number of changepoints in configuration \eqn{C_{1}} and
\eqn{k} is the number of changepoints in configuration \eqn{C_{2}}.
The term \eqn{min(A(C_{1}, C_{2}))} reflects the cost that matches
changepoint locations between \eqn{C_{1}} and \eqn{C_{2}} and cloud be calculated
by linear assignment method. Details can be found in Shi et al. (2022).
Note: if one configuration doesn't contain any changepoint (valued \code{NULL}),
the distance is defined as the \eqn{|m-k|}.
}
\examples{
n = 100

# both tau1 and tau2 has detected changepoints
tau2 = c(25, 50, 75)
tau1 = c(20, 35, 70, 80, 90)
cptDist(tau1=tau1, tau2=tau2, n=n)

# either tau1 or tau2 has zero detected changepoints
cptDist(tau1=tau1, tau2=NULL, n=n)
cptDist(tau1=NULL, tau2=tau2, n=n)
cptDist(tau1=NULL, tau2=NULL, n=n)
}
\references{
{
  Shi, X., Gallagher, C., Lund, R., & Killick, R. (2022).
  A comparison of single and multiple changepoint techniques for time series data.
  \emph{Computational Statistics & Data Analysis}, 170, 107433.
}
}
