% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IslandGA.R
\name{IslandGA}
\alias{IslandGA}
\title{Island model based genetic algorithm}
\usage{
IslandGA(ObjFunc, n, IslandGA_param, IslandGA_operators, ...)
}
\arguments{
\item{ObjFunc}{The fitness function to be minimized. Users can specify any R or Rcpp
functions as the fitness function with setting input as potential solution to
the optimization problem and returning a numerical value as the output/fitness.}

\item{n}{The sample size of the time series.}

\item{IslandGA_param}{A list contains the hyper-parameters for IslandGA.
See \code{\link{IslandGA_param}} for the details.}

\item{IslandGA_operators}{A list includes the functions for population initialization,
new individual selection, and genetic operator of crossover and mutation.
See \code{\link{IslandGA_operators}} for the details.}

\item{...}{additional arguments that will be passed to the fitness function.}
}
\value{
Returns a list that has the following components.
\item{bestfit}{The obtained minimum value of the objective function after
the final iteration.}
\item{bestchrom}{The locations of the detected changepoints associating
with the \code{bestfit} the after the final iteration.}
\item{countMig}{The number of migrations undertaken by the IslandGA.}
\item{convg}{An integer code indicate convergence.
\itemize{
 \item{0} indicates the algorithm successful completion.
 \item{1} indicates the the total number of migrations exceeds the
 pre-specified \item{\code{maxMig}} limit.
 }
}
}
\description{
Perform the modified island-based genetic algorithm (IslandGA) for multiple changepoint detection.
Minimization of an objective function using genetic algorithm (GA).
The algorithm can be run sequentially or in explicit parallelisation.
}
\examples{
library(changepointGA)

Ts = 1000
Cp.prop = c(1/4, 3/4)
CpLocT = floor(Ts*Cp.prop)
DeltaT = c(2, -2)
sigmaT = 1
thetaT = c(0.5) # intercept
XMatT = matrix(1, nrow=Ts, ncol=1)
colnames(XMatT) = "intercept"
myts = ts.sim(theta=thetaT, XMat=XMatT, sigma=sigmaT, Delta=DeltaT, CpLoc=CpLocT)

IslandGA_param = list(
  popsize      = 40,
  Islandsize   = 5,
  Pcrossover   = 0.95,
  Pmutation    = 0.15,
  Pchangepoint = 0.06,
  minDist      = 2,
  mmax         = Ts/2 - 1,
  lmax         = 2 + Ts/2 - 1,
  maxMig       = 500,
  maxgen       = 100,
  maxconv      = 100,
  monitoring   = FALSE,
  parallel     = FALSE, ###
  nCore        = NULL,
  tol          = 1e-5,
  seed         = NULL
)

IslandGA_operators = list(population = "random_population_cpp",
                          selection  = "selection_linearrank_cpp",
                          crossover  = "offspring_uniformcrossover_cpp",
                          mutation   = "mutation")

IslandGA.res = IslandGA(ObjFunc=BinSearch.BIC, n=Ts, IslandGA_param, IslandGA_operators, Xt=myts)
IslandGA.res$bestfit
IslandGA.res$bestchrom
}
