% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{selection_linearrank_cpp}
\alias{selection_linearrank_cpp}
\title{The selection genetic algorithm oeprator}
\usage{
selection_linearrank_cpp(pop, popFit)
}
\arguments{
\item{pop}{A matrix contains the chromosomes for all individuals. The number of
rows is equal to \code{lmax} and the number of columns is equal to the
\code{popsize}.}

\item{popFit}{A vector contains the objective function value (population fit)
being associated to each individual chromosome from above.}
}
\value{
A list contains the chromosomes for dad and mom.
}
\description{
Select a pair of individuals for the crossover operator to produce offspring
(individual for next generation). According to each individual's fittness in
\code{popFit}, the objective function values for each changepoint
configuration, the pair of individuals are selected via the linear rank method.
Among the pair of individuals, we assume that dad has better fit (smaller
objective value or larger rank) than mom.
}
