% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasimulation.R
\name{ts.sim}
\alias{ts.sim}
\title{Time series simulation with changepoint effects}
\usage{
ts.sim(theta, XMat, sigma, phi = NULL, Delta = NULL, CpLoc = NULL, seed = NULL)
}
\arguments{
\item{theta}{A parameter vector contains other mean function parameters without changepoint parameters.}

\item{XMat}{The covairates for time series mean function without changepoint indicators.}

\item{sigma}{The standard deviation for time series residuals \eqn{\epsilon_{t}}.}

\item{phi}{A single value for the autocorrelation parameter for AR(1) errors.}

\item{Delta}{The parameter vector contains the changepoint parameters for time series mean function.}

\item{CpLoc}{A vector contains the changepoint locations range from \eqn{1\leq\tau\leq T_{s}}.}

\item{seed}{The random seed for simulation reproducibility.}
}
\value{
Returns the simulated time series with attributes:
\item{\code{Z}}{The simulated time series.}
\item{Attributes}{
\itemize{
 \item{\code{DesignX}} The covariates include all changepoint indicators.
 \item{\code{mu}} A vector includes the mean values for simulated time series sequences.
 \item{\code{theta}} The true parameter vector (including changepoint effects).
 \item{\code{CpLoc}} The true changepoint locations where we introduce the mean changing effects.
 \item{\code{seed}} The random seed used for this simulation.
 }
}
}
\description{
This is a function to simulate time series \eqn{Z_{t}, t=1,\ldots,T_{s}} from a class of models,
 \deqn{Z_{t}=\mu_{t}+e_{t}.}
\itemize{
 \item{Stationary time series without autocorrelation} \cr \eqn{\mu_{t}} is a constant and \eqn{e_{t}}'s are independent and identically distributed as \eqn{N(0,\sigma)}.
 \item{Stationary time series with autocorrelation} \cr \eqn{\mu_{t}} is a constant
 and \eqn{e_{t}} follows an AR(1) process,
 \deqn{e_{t} = \phi_{t-1}e_{t-1}+\epsilon_{t},}
where \eqn{\epsilon_{t}}'s are independent and identically distributed as \eqn{N(0,\sigma)}.
 \item{Stationary with seasonality and autocorrelation} \cr \eqn{\mu_{t}} is not a constant,
 \deqn{\mu_{t} = ASin(\frac{2\pi t}{S})+BSin(\frac{2\pi t}{S}),}
 and \eqn{e_{t}} follows an AR(1) process,
 \deqn{e_{t} = \phi_{t-1}e_{t-1}+\epsilon_{t},}
where \eqn{\epsilon_{t}}'s are independent and identically distributed as \eqn{N(0,\sigma)}.
 \item{Stationary with seasonality, trend, and autocorrelation} \cr \eqn{\mu_{t}} is not a constant,
 \deqn{\mu_{t} = ASin(\frac{2\pi t}{S})+BSin(\frac{2\pi t}{S}) + \alpha t,}
 and \eqn{e_{t}} follows an AR(1) process,
 \deqn{e_{t} = \phi_{t-1}e_{t-1}+\epsilon_{t},}
where \eqn{\epsilon_{t}}'s are independent and identically distributed as \eqn{N(0,\sigma)}.
}
The changepoint effects could be introduced through the \eqn{\mu_{t}} as
 \deqn{\mu_{t} = \Delta_{1}I_{t>\tau_{1}} + \ldots + \Delta_{m}I_{t>\tau_{m}},}
where \eqn{1\leq\tau_{1}<\ldots<\tau_{m}\leq T_{s}} are the changepoint locations and
\eqn{\Delta_{1},\ldots,\Delta_{m}} are the changepoint parameter that need to be estimated.
}
\examples{
Ts = 1000
Cp.prop = c(1/4, 3/4)
CpLocT = floor(Ts*Cp.prop)
DeltaT = c(2, -2)
sigmaT = 1

##### M1: Stationary time series without autocorrelation
thetaT = c(0.5) # intercept
XMatT = matrix(1, nrow=Ts, ncol=1)
colnames(XMatT) = "intercept"
myts = ts.sim(theta=thetaT, XMat=XMatT, sigma=sigmaT, Delta=DeltaT, CpLoc=CpLocT)
TsPlotCheck(myts, tau=CpLocT)

##### M2: Stationary time series with autocorrelation
phiT = 0.5
thetaT = c(0.5) # intercept
XMatT = matrix(1, nrow=Ts, ncol=1)
colnames(XMatT) = "intercept"
myts = ts.sim(theta=thetaT, XMat=XMatT, phi=phiT, sigma=sigmaT, Delta=DeltaT, CpLoc=CpLocT)
TsPlotCheck(myts, tau=CpLocT)

##### M3: Stationary with seasonality and autocorrelation
thetaT = c(0.5, -0.5, 0.3) # intercept, B, D
period = 30
XMatT = cbind(rep(1, Ts), cos(2*pi*(1:Ts)/period), sin(2*pi*(1:Ts)/period))
colnames(XMatT) = c("intercept", "Bvalue", "DValue")
myts = ts.sim(theta=thetaT, XMat=XMatT, phi=phiT, sigma=sigmaT, Delta=DeltaT, CpLoc=CpLocT)
TsPlotCheck(myts, tau=CpLocT)

##### M4: Stationary with seasonality, trend, and autocorrelation
# scaled trend if large number of sample size
thetaT = c(0.5, -0.5, 0.3, 0.01) # intercept, B, D, alpha
period = 30
XMatT = cbind(rep(1, Ts), cos(2*pi*(1:Ts)/period), sin(2*pi*(1:Ts)/period), 1:Ts)
colnames(XMatT) = c("intercept", "Bvalue", "DValue", "trend")
myts = ts.sim(theta=thetaT, XMat=XMatT, phi=phiT, sigma=sigmaT, Delta=DeltaT, CpLoc=CpLocT)
TsPlotCheck(myts, tau=CpLocT)
}
