% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasimulation.R
\name{TsPlotCheck}
\alias{TsPlotCheck}
\title{Plot the simulated time series}
\usage{
TsPlotCheck(Z, tau = NULL, mu = NULL)
}
\arguments{
\item{Z}{The simulated time series.}

\item{tau}{The provided changepoint locations.}

\item{mu}{The provided meam values for each time \eqn{t}.}
}
\description{
This is a function to plot the simulated time series with segmentation visualization
by provided changepoint locations.
}
\examples{
##### M1: Stationary time series without autocorrelation
Ts = 1000
Cp.prop = c(1/4, 3/4)
CpLocT = floor(Ts*Cp.prop)
DeltaT = c(2, -2)
sigmaT = 1
thetaT = c(0.5) # intercept
XMatT = matrix(1, nrow=Ts, ncol=1)
colnames(XMatT) = "intercept"
myts = ts.sim(theta=thetaT, XMat=XMatT, sigma=sigmaT, Delta=DeltaT, CpLoc=CpLocT)

TsPlotCheck(myts, tau=CpLocT)
}
