% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlantaTemperature.R
\docType{data}
\name{AtlantaTemperature}
\alias{AtlantaTemperature}
\title{Atlanta airport yearly average temperatures data}
\format{
An object of class \code{data.frame} with 134 rows and 2 columns.
}
\usage{
data(AtlantaTemperature)
}
\description{
The annual mean temperature data from 1879 to 2013 at Atlanta, Georgia’s
Hartsfield International Airport weather station.
}
\examples{
data(AtlantaTemperature)
}
\references{
Lund, R. B., Beaulieu, C., Killick, R., Lu, Q., & Shi, X. (2023).
Good practices and common pitfalls in climate time series changepoint techniques:
A review. \emph{Journal of Climate}, 36(23), 8041-8057.
}
\keyword{datasets}
