% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAfunc.R
\name{BinSearch.BIC}
\alias{BinSearch.BIC}
\title{The example objective function for changepoint search in First order Gaussian
autoregressions (AR(1)) via Bayesian Information Criterion (BIC)}
\usage{
BinSearch.BIC(tau, m, Xt)
}
\arguments{
\item{tau}{The provided changepoint locations (Only locations are included).}

\item{m}{The provided number of changepoint locations.}

\item{Xt}{The simulated AR(1) time series from \code{ts.sim} function.}
}
\value{
Returned the value of the obejctive function (i.e. BIC).
}
\description{
This is a function to calculate
}
\examples{
library(changepointGA)

Ts = 1000
Cp.prop = c(1/4, 3/4)
CpLocT = floor(Ts*Cp.prop)
DeltaT = c(2, -2)
sigmaT = 1
thetaT = c(0.5) # intercept
XMatT = matrix(1, nrow=Ts, ncol=1)
colnames(XMatT) = "intercept"
myts = ts.sim(theta=thetaT, XMat=XMatT, sigma=sigmaT, Delta=DeltaT, CpLoc=CpLocT)

# candidate changepoint configuration
tautest = c(250, 500, 750)
BinSearch.BIC(tau=tautest, m=3, Xt=myts)
}
