% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GA.R
\name{GA}
\alias{GA}
\title{Genetic algorithm}
\usage{
GA(ObjFunc, n, GA_param, ga_operators, ...)
}
\arguments{
\item{ObjFunc}{The fitness function to be minimized. Users can specify any R or Rcpp
functions as the fitness function with setting input as potential solution to
the optimization problem and returning a numerical value as the output/fitness.}

\item{n}{The sample size of the time series.}

\item{GA_param}{A list contains the hyper-parameters for genetic algorithm.
See \code{\link{GA_param}} for the details.}

\item{ga_operators}{A list includes the functions for population initialization,
new individual selection, and genetic operator of crossover and mutation.
See \code{\link{ga_operators}} for the details.}

\item{...}{additional arguments that will be passed to the fitness function.}
}
\value{
Returns a list that has components:
\item{overbestfit}{The obtained minimum value of the objective function after
the final iteration.}
\item{overbestchrom}{The locations of the detected changepoints associating
with the \code{overbestfit} the after the final iteration.}
\item{bestfit}{The minimized fitness function values at each iteration.}
\item{bestchrom}{The detected changepoints at each iteration.}
\item{count}{The number of iterations undertaken by the genetic algorithm.}
\item{convg}{An integer code.
\itemize{
 \item{0} indicates the algorithm successful completion.
 \item{1} indicates the the total number of generations exceeds the
 prespecified \item{\code{maxgen}} limit.
 }
}
}
\description{
Perform the modified genetic algorithm for multiple changepoint detection.
Minimization of an objective function using genetic algorithm (GA).
The algorithm can be run sequentially or in explicit parallelisation.
}
\examples{
library(changepointGA)

Ts = 1000
Cp.prop = c(1/4, 3/4)
CpLocT = floor(Ts*Cp.prop)
DeltaT = c(2, -2)

sigmaT = 1

thetaT = c(0.5) # intercept

XMatT = matrix(1, nrow=Ts, ncol=1)
colnames(XMatT) = "intercept"
myts = ts.sim(theta=thetaT, XMat=XMatT, sigma=sigmaT, Delta=DeltaT, CpLoc=CpLocT)

GA_param = list(
  popsize      = 200,
  Pcrossover   = 0.95,
  Pmutation    = 0.15,
  Pchangepoint = 0.06,
  minDist      = 2,
  mmax         = Ts/2 - 1,
  lmax         = 2 + Ts/2 - 1,
  maxgen       = 10000,
  maxconv      = 10000,
  monitoring   = FALSE,
  parallel     = FALSE,
  nCore        = NULL,
  tol          = 1e-5,
  seed         = NULL
)

ga_operators = list(population = "random_population_cpp",
                    selection  = "selection_linearrank_cpp",
                    crossover  = "offspring_uniformcrossover_cpp",
                    mutation   = "mutation")

GA.res = GA(BinSearch.BIC, n=Ts, GA_param, ga_operators, Xt=myts)
GA.res$overbestfit
GA.res$overbestchrom
}
