% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IslandGA.R
\name{IslandGA}
\alias{IslandGA}
\title{Island model based genetic algorithm}
\usage{
IslandGA(
  ObjFunc,
  N,
  IslandGA_param = .default.IslandGA_param,
  IslandGA_operators = .default.IslandGA_operators,
  p.range = NULL,
  ...
)
}
\arguments{
\item{ObjFunc}{The fitness function to be minimized. Users can specify any R or Rcpp
functions as the fitness function with setting input as potential solution to
the optimization problem and returning a numerical value as the output/fitness.}

\item{N}{The sample size of the time series.}

\item{IslandGA_param}{A list contains the hyper-parameters for IslandGA.
See \code{\link{IslandGA_param}} for the details.}

\item{IslandGA_operators}{A list includes the functions for population initialization,
new individual selection, and genetic operator of crossover and mutation.
See \code{\link{IslandGA_operators}} for the details.}

\item{p.range}{Default is \code{NULL} for only changepoint detection. If
\code{p.range} is specified as a list object, which contains the range of
each model order parameters for order selection (integers). The number of
order parameters must be equal to the length of \code{p.range}.}

\item{...}{additional arguments that will be passed to the fitness function.}
}
\value{
Returns a list that has the following components.
\item{overbestfit}{The obtained minimum value of the objective function after
the final iteration.}
\item{overbestchrom}{The locations of the detected changepoints associating
with the \code{overbestfit} the after the final iteration.}
\item{bestfit}{The minimized fitness function values at each iteration.}
\item{bestchrom}{The detected changepoints at each iteration.}
\item{countMig}{The number of migrations undertaken by the IslandGA.}
\item{count}{The number of iterations (generations) undertaken by the island
genetic algorithm model.}
\item{convg}{An integer code indicate convergence.}
\itemize{
 \item{0} indicates the algorithm successful completion.
 \item{1} indicates the the total number of migrations exceeds the
 pre-specified \item{\code{maxMig}} limit.
 }
}
\description{
Perform the modified island-based genetic algorithm (IslandGA) for multiple changepoint detection.
Minimization of an objective function using genetic algorithm (GA).
The algorithm can be run sequentially or in explicit parallelisation.
}
\examples{
Ts = 1000
betaT = c(0.5) # intercept
XMatT = matrix(1, nrow=Ts, ncol=1)
colnames(XMatT) = "intercept"
sigmaT = 1
phiT = c(0.5, -0.5)
thetaT = c(0.8)
DeltaT = c(2, -2)
Cp.prop = c(1/4, 3/4)
CpLocT = floor(Ts*Cp.prop)

myts = ts.sim(beta=betaT, XMat=XMatT, sigma=sigmaT, phi=phiT, theta=thetaT,
              Delta=DeltaT, CpLoc=CpLocT, seed=1234)
TsPlotCheck(Y=myts, tau=CpLocT)

IslandGA.res = IslandGA(ObjFunc=BinSearch.BIC, N=Ts, Xt=myts)
# IslandGA.res$overbestfit
# IslandGA.res$overbestchrom
}
